Fix bugs (#105)
* fix role change in user update

* fix category update dto

* fix not allowed to update admin info

* fix not allowed to delete admin

* fix (Graph) :edge and node upload faster now

* fix (Test) : fix some test

+++ AnalysisData/AnalysisData/Services/UserService/AdminService/AdminService.cs
@@ -29,7 +29,10 @@ public class AdminService : IAdminService
     public async Task UpdateUserInformationByAdminAsync(Guid id, UpdateAdminDto updateAdminDto)
     {
         var user = await _userRepository.GetUserByIdAsync(id);
-        
+        if (user !=null && user.Username == "admin")
+        {
+            throw new AdminProtectedException();
+        }
         await ValidateUserInformation(user, updateAdminDto);
         _validationService.EmailCheck(updateAdminDto.Email);
         _validationService.PhoneNumberCheck(updateAdminDto.PhoneNumber);
@@ -70,13 +73,18 @@ public class AdminService : IAdminService
         {
             throw new RoleNotFoundException();
         }
-        user.Role = role;
+        user.RoleId = role.Id;
         await _userRepository.UpdateUserAsync(user.Id, user);
     }
     
 
     public async Task<bool> DeleteUserAsync(Guid id)
     {
+        var user = await _userRepository.GetUserByIdAsync(id);
+        if (user != null && user.Username == "admin")
+        {
+            throw new AdminProtectedException();
+        }
         var isDelete = await _userRepository.DeleteUserAsync(id);
         if (!isDelete)
             throw new UserNotFoundException();
